requirements.txt
pandas
django
djangorestframework
django-bootstrap4